pyproject.toml
[project]
name = "rx-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "routeros-api>=0.21.0",
    "sentry-sdk[fastapi]>=2.39.0",
]

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.3",
]


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "devtools", "dotenv", "routeros_api", "sentry_sdk"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
